LPS-56552 Move SafeFileNameStoreWrapper and StoreProxyImpl to its module
