+++ modules/portal/portal-store-safefilenamewrapper/src/com/liferay/portal/store/safefilenamewrapper/SafeFileNameStoreWrapper.java
@@ -12,11 +12,12 @@
  * details.
  */
 
-package com.liferay.portlet.documentlibrary.store;
+package com.liferay.portal.store.safefilenamewrapper;
 
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.util.FileUtil;
 import com.liferay.portlet.documentlibrary.model.DLFileEntryConstants;
+import com.liferay.portlet.documentlibrary.store.Store;
 
 import java.io.File;
 import java.io.InputStream;
