+++ portal-impl/src/META-INF/util-spring.xml
@@ -842,11 +842,6 @@
 	<bean id="com.liferay.portlet.documentlibrary.service.permission.DLFolderPermission" class="com.liferay.portlet.documentlibrary.service.permission.DLFolderPermission" />
 	<bean id="com.liferay.portlet.documentlibrary.service.permission.DLPermission" class="com.liferay.portlet.documentlibrary.service.permission.DLPermission" />
 	<bean id="com.liferay.portlet.documentlibrary.store.DLStore" class="com.liferay.portlet.documentlibrary.store.DLStoreImpl" />
-	<bean id="com.liferay.portlet.documentlibrary.store.Store" class="com.liferay.portlet.documentlibrary.store.SafeFileNameStoreWrapper">
-		<constructor-arg>
-			<bean class="com.liferay.portlet.documentlibrary.store.StoreProxyImpl" />
-		</constructor-arg>
-	</bean>
 	<bean id="com.liferay.portlet.documentlibrary.store.StoreFactory" class="com.liferay.portlet.documentlibrary.store.StoreFactory" factory-method="getInstance" />
 	<bean id="com.liferay.portlet.documentlibrary.util.DLFileEntryIndexer" class="com.liferay.portlet.documentlibrary.util.DLFileEntryIndexer" />
 	<bean id="com.liferay.portlet.documentlibrary.util.DLFileEntryLockListener" class="com.liferay.portlet.documentlibrary.util.DLFileEntryLockListener" />
